+++ core/src/com/hro/hrogame/gameobject/effect/residualeffect/FreezeOverTimeEffect.java
@@ -114,7 +114,7 @@ public class FreezeOverTimeEffect extends Effect {
         return data.duration.current;
     }
     @Override
-    public int getEffectWeight() {
+    public float getEffectWeight() {
         return 0;
     }
     // endregion
