+++ core/src/com/hro/hrogame/gameobject/effect/residualeffect/ShieldOverTimeEffect.java
@@ -111,7 +111,7 @@ public class ShieldOverTimeEffect extends Effect {
         return data.duration.current;
     }
     @Override
-    public int getEffectWeight() {
+    public float getEffectWeight() {
         return 0;
     }
     // endregion
