+++ core/src/com/hro/hrogame/gameobject/effect/selfdestruction/SelfDestructionEffect.java
@@ -58,7 +58,7 @@ public class SelfDestructionEffect extends Effect {
         return 0;
     }
     @Override
-    public int getEffectWeight() {
+    public float getEffectWeight() {
         return data.weight;
     }
     // endregion
