+++ core/src/com/hro/hrogame/gameobject/effect/waveeffect/StunnerEffect.java
@@ -97,7 +97,7 @@ public class StunnerEffect extends Effect {
         return data.cooldown.current;
     }
     @Override
-    public int getEffectWeight() {
+    public float getEffectWeight() {
         return data.weight;
     }
     // endregion
